requirements.txt
# ============================================
# StockMind + JyotishMarket AI
# All Dependencies
# ============================================

# ----- Core AI Framework -----
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# ----- LLM Provider -----
together>=0.2.0

# ----- Stock Market Data -----
yfinance>=0.2.33

# ----- News & Sentiment -----
newsapi-python>=0.2.7

# ----- Web Interface -----
streamlit>=1.30.0

# ----- Data Processing -----
pandas>=2.1.4
plotly>=5.18.0

# ----- Vector Store (for future RAG features) -----
chromadb>=0.4.22

# ----- Vedic Astrology -----
kerykeion>=4.0.0

# ----- Environment Management -----
python-dotenv>=1.0.0

# ----- Utilities -----
tiktoken>=0.5.2